+++ E2E/.gitignore
@@ -1,4 +1,3 @@
-node_modules/*
-!node_modules/config
-!node_modules/helpers
-storage/
+**node_modules**
+src/storage/
+src/test-results/
