+++ E2E/config/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "@lepine/e2e-config",
+    "version": "1.0.0"
+}
