+++ E2E/helpers/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "@lepine/e2e-helpers",
+    "version": "1.0.0"
+}
