+++ E2E/package.json
@@ -1,22 +1,8 @@
 {
-    "name": "E2E",
-    "version": "1.0.0",
-    "description": "E2E tests for LePine transfer system",
-    "main": "index.js",
-    "author": "Fube",
-    "license": "BSD-3-Clause",
-    "devDependencies": {
-        "@playwright/test": "^1.17.1",
-        "axios": "^0.24.0",
-        "nanoid": "^3.1.30"
-    },
-    "scripts": {
-        "test": "npx playwright test",
-        "test-debug": "PWDEBUG=1 npx playwright test",
-        "test-debug-windows": "set PWDEBUG=1 && npx playwright test"
-    },
     "private": true,
     "workspaces": [
-        "config"
+        "config",
+        "helpers",
+        "src"
     ]
 }
