+++ E2E/src/tests/auth/logout.spec.js
